+++ src/app/auth/sign-in/page.tsx
@@ -1,0 +1,7 @@
+import SignInAuthForm from "@/features/auth/sing-in/sign-in-auth-form";
+
+const SignInPage = () => {
+  return <SignInAuthForm />;
+};
+
+export default SignInPage;
