+++ src/app/auth/sign-up/character/page.tsx
@@ -1,0 +1,5 @@
+const SignUpCharacterPage = () => {
+  return <div>character를 선택해주세요</div>;
+};
+
+export default SignUpCharacterPage;
