+++ src/app/auth/sign-up/page.tsx
@@ -1,0 +1,7 @@
+import SignUpAuthForm from '@/features/auth/sign-up/sign-up-auth-form';
+
+const SignUpPage = () => {
+  return <SignUpAuthForm />;
+};
+
+export default SignUpPage;
