--- src/app/sign-in/page.tsx
@@ -1,7 +1,0 @@
-import AuthForm from '@/components/common/auth-form';
-
-const LoginPage = () => {
-  return <AuthForm />;
-};
-
-export default LoginPage;
