--- src/app/sign-up/page.tsx
@@ -1,7 +1,0 @@
-import AuthForm from '@/components/common/auth-form';
-
-const SignupPage = () => {
-  return <AuthForm />;
-};
-
-export default SignupPage;
