+++ src/constants/path-constant.ts
@@ -1,7 +1,15 @@
 export const PATH = {
   MAIN: '/',
-  SIGN_IN: '/sign-in',
-  SIGN_UP: '/sign-up',
+  AUTH: {
+    SIGN_IN: '/auth/sign-in',
+    SIGN_UP: '/auth/sign-up',
+    CHARACTER: '/auth/character',
+  },
   ON_BOARDING: '/on-boarding',
-  MY_PAGE: '/my-page',
+};
+
+export const ROUTE_HANDLER_PATH = {
+  AUTH: {
+    SIGN_UP: '/api/sign-up',
+  },
 };
